Moved FirstJoinTask method bodies out of AnalysisPipeline
* RunFdr, RunPercolatorFdr, BuildBasicFeatures, RunPercolatorStreaming,
  RunSimpleFdr, RunFirstPassProteinFdr, WriteFdrScoresSidecars,
  WriteReconciliationFiles, ResolveSidecarBasePath, and
  BuildReconciliationFile become private members of FirstJoinTask
  (10 methods, ~810 lines)
* LogInfo / LogWarning / LogError calls route through _ctx.LogX
  where _ctx is the PipelineContext set on Run entry. AnalysisPipeline
  back-reference dropped from the constructor
* NUM_PIN_FEATURES const promoted from private to internal so the
  moved code can keep using it (qualified as
  AnalysisPipeline.NUM_PIN_FEATURES)
* AnalysisPipeline.cs shrinks by another ~810 lines
* Stellar 3-file snapshot regression PASS at every stage

See ai/todos/active/TODO-20260509_osprey_pipeline_tasks.md
