+++ pwiz_tools/OspreySharp/OspreySharp/AnalysisPipeline.cs
@@ -51,7 +51,10 @@ namespace pwiz.OspreySharp
     /// </summary>
     public partial class AnalysisPipeline
     {
-        private const int NUM_PIN_FEATURES = 21;
+        // Internal so FirstJoinTask (which now owns RunPercolatorFdr +
+        // RunPercolatorStreaming + BuildBasicFeatures) can reuse the
+        // same 21-feature width without redeclaring it.
+        internal const int NUM_PIN_FEATURES = 21;
 
         // EntryId encodes target/decoy in the high bit; base_id is the
         // lower 31 bits, shared by a target and its paired decoy.
@@ -146,7 +149,7 @@ namespace pwiz.OspreySharp
                 // exit; in either case ctx.ExitCode is the process
                 // exit code AnalysisPipeline.Run should propagate.
                 var firstJoinTask = new FirstJoin.FirstJoinTask(
-                    this, perFileEntries, perFileCalibrations,
+                    perFileEntries, perFileCalibrations,
                     perFileParquetPaths, fullLibrary);
                 var firstJoinCtx = new Tasks.PipelineContext(
                     config, LogInfo, LogWarning, LogError);
@@ -705,343 +708,6 @@ namespace pwiz.OspreySharp
             }
 
             return scoredEntries;
-        }
-
-        /// <summary>
-        /// Write the per-file <c>.1st-pass.fdr_scores.bin</c> sidecars at
-        /// the pre-compaction Stage 5 boundary (every stub, passing or
-        /// not, gets persisted with its q-values + SVM score). Mirrors
-        /// the persist_fdr_scores call in osprey/src/pipeline.rs at line
-        /// ~3180 (immediately after first-pass FDR, before compaction or
-        /// protein FDR). Stage 6 workers re-apply the q-value threshold
-        /// themselves to derive the post-compaction passing set.
-        /// </summary>
-        /// <returns>
-        /// Number of files for which the sidecar write failed (0 means
-        /// success). Callers in <see cref="OspreyConfig.StopAfterStage5"/>
-        /// mode treat any failure as fatal — see the StopAfterStage5
-        /// block at the end of the reconciliation phase — because the
-        /// downstream Stage 6 worker would otherwise be missing a
-        /// sidecar.
-        /// </returns>
-        internal int WriteFdrScoresSidecars(
-            List<KeyValuePair<string, List<FdrEntry>>> perFileEntries,
-            Dictionary<string, string> perFileParquetPaths,
-            OspreyConfig config)
-        {
-            int failures = 0;
-            foreach (var kvp in perFileEntries)
-            {
-                string fileName = kvp.Key;
-                string sidecarBase = ResolveSidecarBasePath(fileName, perFileParquetPaths, config);
-                if (string.IsNullOrEmpty(sidecarBase))
-                {
-                    LogWarning(string.Format(
-                        "No sidecar base path for `{0}` — skipping fdr_scores.bin write", fileName));
-                    failures++;
-                    continue;
-                }
-                string fdrPath = FdrScoresSidecar.Pass1Path(sidecarBase);
-                try
-                {
-                    FdrScoresSidecar.Write(fdrPath, kvp.Value, FdrScoresSidecar.Pass.FirstPass);
-                }
-                catch (Exception ex)
-                {
-                    LogWarning(string.Format(
-                        "Failed to write 1st-pass fdr_scores.bin for {0}: {1}", fileName, ex.Message));
-                    failures++;
-                }
-            }
-            return failures;
-        }
-
-        /// <summary>
-        /// Write the per-file <c>.reconciliation.json</c> envelope for
-        /// each input — the second half of the Stage 5 → Stage 6 boundary
-        /// (the <c>.fdr_scores.bin</c> companion is written earlier,
-        /// pre-compaction). Mirrors the matching block in
-        /// osprey/src/pipeline.rs immediately after
-        /// dump_stage6_reconciliation. The file is written sibling to
-        /// the input mzML (or, in --join-only mode, the synthetic input
-        /// path derived from the parquet stem).
-        /// </summary>
-        /// <returns>
-        /// Number of files for which the reconciliation.json write failed
-        /// (0 means success). Callers in
-        /// <see cref="OspreyConfig.StopAfterStage5"/> mode treat any
-        /// failure as fatal — the Stage 6 worker would otherwise be
-        /// missing an envelope and either refuse to start or score the
-        /// wrong files.
-        /// </returns>
-        internal int WriteReconciliationFiles(
-            List<KeyValuePair<string, List<FdrEntry>>> perFileEntries,
-            IReadOnlyDictionary<(string File, int Index), ReconcileAction> reconciliationActions,
-            IReadOnlyList<PeptideConsensusRT> consensus,
-            Dictionary<string, RTCalibration> refinedCalibrations,
-            IReadOnlyDictionary<string, RTCalibration> perFileCalibrations,
-            List<LibraryEntry> fullLibrary,
-            Dictionary<string, string> perFileParquetPaths,
-            OspreyConfig config,
-            out Dictionary<string, List<GapFillTarget>> gapFillByFileOut)
-        {
-            string searchHash = config.SearchParameterHash();
-            string libraryHash = config.LibraryIdentityHash();
-            var actions = reconciliationActions
-                ?? new Dictionary<(string File, int Index), ReconcileAction>();
-
-            // Pre-group reconciliation actions by file name to avoid the
-            // O(num_files * num_actions) walk that the previous
-            // implementation performed inside BuildReconciliationFile
-            // (one full Dictionary traversal per file).
-            var actionsByFile = new Dictionary<string, List<KeyValuePair<int, ReconcileAction>>>(
-                StringComparer.Ordinal);
-            foreach (var kvp in actions)
-            {
-                if (!actionsByFile.TryGetValue(kvp.Key.File, out var list))
-                {
-                    list = new List<KeyValuePair<int, ReconcileAction>>();
-                    actionsByFile[kvp.Key.File] = list;
-                }
-                list.Add(new KeyValuePair<int, ReconcileAction>(kvp.Key.Index, kvp.Value));
-            }
-
-            // Build the (modified_sequence, charge) → (target_id, decoy_id)
-            // and entry_id → precursor_mz lookups from the library. Decoy
-            // ID convention: target_id | 0x80000000 (mirrors Rust at
-            // pipeline.rs:3330-3340).
-            var libLookup = new Dictionary<(string ModifiedSequence, byte Charge),
-                (uint TargetEntryId, uint DecoyEntryId)>();
-            var libPrecursorMz = new Dictionary<uint, double>();
-            foreach (var entry in fullLibrary)
-            {
-                if (entry.IsDecoy)
-                    continue;
-                uint decoyId = entry.Id | 0x80000000u;
-                libLookup[(entry.ModifiedSequence, entry.Charge)] = (entry.Id, decoyId);
-                libPrecursorMz[entry.Id] = entry.PrecursorMz;
-            }
-
-            // Compute per-file gap-fill targets. Per-file isolation-window
-            // m/z intervals are not yet plumbed through C# (Stellar
-            // calibration.json carries no isolation_scheme today, so the
-            // filter is a no-op there); when extended to GPF datasets,
-            // pass a non-null dictionary here.
-            var perFileForGapFill = new List<KeyValuePair<string,
-                IReadOnlyList<FdrEntry>>>(perFileEntries.Count);
-            foreach (var kvp in perFileEntries)
-            {
-                perFileForGapFill.Add(new KeyValuePair<string,
-                    IReadOnlyList<FdrEntry>>(kvp.Key, kvp.Value));
-            }
-            var gapFillByFile = GapFillTargetIdentifier.Identify(
-                consensus,
-                perFileForGapFill,
-                refinedCalibrations,
-                perFileCalibrations,
-                config.Reconciliation.ConsensusFdr,
-                libLookup,
-                libPrecursorMz,
-                perFileIsolationMz: null);
-
-            // Mirror gapFillByFile into the out param for the in-process
-            // Stage 6 rescore caller. Identify returns IReadOnlyList<>;
-            // ExecuteStage6Rescore's perFileGapFill type is
-            // IReadOnlyDictionary<string, List<GapFillTarget>>, so build
-            // a List<> per file. The conversion is per-file (3-15 files
-            // typical), each list 100-3000 GapFillTarget structs.
-            gapFillByFileOut = new Dictionary<string, List<GapFillTarget>>(
-                gapFillByFile.Count, StringComparer.Ordinal);
-            foreach (var kvp in gapFillByFile)
-            {
-                var copy = new List<GapFillTarget>(kvp.Value.Count);
-                foreach (var g in kvp.Value)
-                    copy.Add(g);
-                gapFillByFileOut[kvp.Key] = copy;
-            }
-
-            int failures = 0;
-            foreach (var kvp in perFileEntries)
-            {
-                string fileName = kvp.Key;
-                var fileEntries = kvp.Value;
-
-                string sidecarBase = ResolveSidecarBasePath(fileName, perFileParquetPaths, config);
-                if (string.IsNullOrEmpty(sidecarBase))
-                {
-                    LogWarning(string.Format(
-                        "No sidecar base path for `{0}` — skipping reconciliation.json write", fileName));
-                    failures++;
-                    continue;
-                }
-                string reconPath = ReconciliationFile.PathForInput(sidecarBase);
-                IReadOnlyList<GapFillTarget> fileGapFill;
-                if (!gapFillByFile.TryGetValue(fileName, out fileGapFill))
-                    fileGapFill = Array.Empty<GapFillTarget>();
-                actionsByFile.TryGetValue(fileName, out var fileActions);
-                var reconFile = BuildReconciliationFile(
-                    fileEntries, fileActions, fileGapFill,
-                    refinedCalibrations.TryGetValue(fileName, out var fileCal) ? fileCal : null,
-                    searchHash, libraryHash);
-                try
-                {
-                    ReconciliationFile.Save(reconPath, reconFile);
-                    LogInfo(string.Format(
-                        "Wrote reconciliation.json for {0} ({1} use_cwt + {2} forced + {3} gap-fill)",
-                        fileName,
-                        reconFile.UseCwtPeakActions.Count,
-                        reconFile.ForcedIntegrationActions.Count,
-                        reconFile.GapFillTargets.Count));
-                }
-                catch (Exception ex)
-                {
-                    LogWarning(string.Format(
-                        "Failed to write reconciliation.json for {0}: {1}", fileName, ex.Message));
-                    failures++;
-                }
-            }
-            return failures;
-        }
-
-        /// <summary>
-        /// Resolve a path whose stem matches <paramref name="fileName"/>, used
-        /// only as the base for sidecar file naming (the path itself need
-        /// not exist). In normal mode this is the input mzML; in
-        /// --join-only mode where InputFiles is empty we synthesize the
-        /// path from the matching .scores.parquet by replacing the
-        /// `.scores.parquet` suffix with `.mzML`. Mirrors the Rust
-        /// `synthetic_input_from_parquet` helper.
-        /// </summary>
-        private static string ResolveSidecarBasePath(
-            string fileName,
-            Dictionary<string, string> perFileParquetPaths,
-            OspreyConfig config)
-        {
-            // Normal mode: prefer the actual input mzML path so sidecars
-            // land next to the source mzML.
-            if (config.InputFiles != null)
-            {
-                foreach (string inputPath in config.InputFiles)
-                {
-                    if (string.Equals(
-                        Path.GetFileNameWithoutExtension(inputPath),
-                        fileName,
-                        StringComparison.Ordinal))
-                    {
-                        return inputPath;
-                    }
-                }
-            }
-            // --join-only fallback: derive a synthetic mzML path from the
-            // matching parquet stem so all the existing sidecar path
-            // helpers keep working without conditional branches.
-            if (perFileParquetPaths != null
-                && perFileParquetPaths.TryGetValue(fileName, out string parquetPath))
-            {
-                string parent = Path.GetDirectoryName(parquetPath) ?? ".";
-                return Path.Combine(parent, fileName + ".mzML");
-            }
-            return null;
-        }
-
-        /// <summary>
-        /// Convert pre-grouped reconciliation actions for one file into
-        /// the <see cref="ReconciliationFile"/> wire format: resolve each
-        /// Vec index to its entry_id, split non-Keep actions into
-        /// homogeneous use_cwt_peak / forced arrays, snapshot the
-        /// refined RT calibration if present, and emit the gap-fill
-        /// targets for this file (already sorted by
-        /// <c>target_entry_id</c> by the identifier). The caller pre-
-        /// groups actions by file so the per-file cost stays
-        /// O(actions_for_this_file) rather than O(total_actions).
-        /// </summary>
-        private static ReconciliationFile BuildReconciliationFile(
-            IReadOnlyList<FdrEntry> fileEntries,
-            IReadOnlyList<KeyValuePair<int, ReconcileAction>> fileActions,
-            IReadOnlyList<GapFillTarget> gapFillTargets,
-            RTCalibration refinedCalibration,
-            string searchHash,
-            string libraryHash)
-        {
-            var useCwt = new List<UseCwtPeakEntry>();
-            var forced = new List<ForcedIntegrationEntry>();
-            if (fileActions != null)
-            {
-                foreach (var kvp in fileActions)
-                {
-                    int idx = kvp.Key;
-                    if (idx < 0 || idx >= fileEntries.Count)
-                        continue;
-                    uint entryId = fileEntries[idx].EntryId;
-                    var useCwtAction = kvp.Value as ReconcileAction.UseCwtPeak;
-                    var forcedAction = kvp.Value as ReconcileAction.ForcedIntegration;
-                    if (useCwtAction != null)
-                    {
-                        useCwt.Add(new UseCwtPeakEntry
-                        {
-                            ApexRt = useCwtAction.ApexRt,
-                            CandidateIdx = (uint)useCwtAction.CandidateIndex,
-                            EndRt = useCwtAction.EndRt,
-                            EntryId = entryId,
-                            StartRt = useCwtAction.StartRt,
-                        });
-                    }
-                    else if (forcedAction != null)
-                    {
-                        forced.Add(new ForcedIntegrationEntry
-                        {
-                            EntryId = entryId,
-                            ExpectedRt = forcedAction.ExpectedRt,
-                            HalfWidth = forcedAction.HalfWidth,
-                        });
-                    }
-                }
-            }
-            // Sort by entry_id for deterministic output (matches Rust).
-            useCwt.Sort((a, b) => a.EntryId.CompareTo(b.EntryId));
-            forced.Sort((a, b) => a.EntryId.CompareTo(b.EntryId));
-
-            RefinedRtCalibrationJson refinedJson = null;
-            if (refinedCalibration != null)
-            {
-                refinedJson = new RefinedRtCalibrationJson
-                {
-                    AbsResiduals = (double[])refinedCalibration.AbsResiduals.Clone(),
-                    FittedRts = (double[])refinedCalibration.FittedValues.Clone(),
-                    LibraryRts = (double[])refinedCalibration.LibraryRts.Clone(),
-                    ResidualSd = refinedCalibration.ResidualSD,
-                };
-            }
-
-            // Map per-file GapFillTarget records (already sorted by
-            // target_entry_id) to the wire form. Field-for-field copy.
-            var gap = new List<GapFillEntry>(gapFillTargets?.Count ?? 0);
-            if (gapFillTargets != null)
-            {
-                foreach (var g in gapFillTargets)
-                {
-                    gap.Add(new GapFillEntry
-                    {
-                        Charge = g.Charge,
-                        DecoyEntryId = g.DecoyEntryId,
-                        ExpectedRt = g.ExpectedRt,
-                        HalfWidth = g.HalfWidth,
-                        ModifiedSequence = g.ModifiedSequence,
-                        TargetEntryId = g.TargetEntryId,
-                    });
-                }
-            }
-
-            return new ReconciliationFile
-            {
-                ForcedIntegrationActions = forced,
-                FormatVersion = ReconciliationFile.CurrentFormatVersion,
-                GapFillTargets = gap,
-                LibraryHash = libraryHash,
-                RefinedRtCalibration = refinedJson,
-                SearchHash = searchHash,
-                UseCwtPeakActions = useCwt,
-            };
         }
 
         /// <summary>
@@ -4901,487 +4567,9 @@ namespace pwiz.OspreySharp
 
         #region Stage 5: FDR Control
 
-        /// <summary>
-        /// Run FDR control using the configured method.
-        /// </summary>
-        internal void RunFdr(
-            List<KeyValuePair<string, List<FdrEntry>>> perFileEntries,
-            List<LibraryEntry> fullLibrary,
-            OspreyConfig config)
-        {
-            switch (config.FdrMethod)
-            {
-                case FdrMethod.Percolator:
-                    RunPercolatorFdr(perFileEntries, fullLibrary, config);
-                    break;
-
-                case FdrMethod.Simple:
-                    RunSimpleFdr(perFileEntries, config);
-                    break;
-
-                default:
-                    LogWarning(string.Format(
-                        "FDR method {0} not yet supported, falling back to simple",
-                        config.FdrMethod));
-                    RunSimpleFdr(perFileEntries, config);
-                    break;
-            }
-        }
-
-        /// <summary>
-        /// Run Percolator-based FDR control.
-        /// Builds PercolatorEntry objects from FdrEntry stubs and runs Percolator.
-        /// </summary>
-        private void RunPercolatorFdr(
-            List<KeyValuePair<string, List<FdrEntry>>> perFileEntries,
-            List<LibraryEntry> fullLibrary,
-            OspreyConfig config)
-        {
-            // Build PercolatorEntry list from all files
-            var percEntries = new List<PercolatorEntry>();
-
-            // Build library lookup for feature extraction
-            var libraryById = new Dictionary<uint, LibraryEntry>(fullLibrary.Count);
-            foreach (var entry in fullLibrary)
-                libraryById[entry.Id] = entry;
-
-            int nWithFeatures = 0;
-            int nWithoutFeatures = 0;
-            int nInputTargets = 0;
-            int nInputDecoys = 0;
-            foreach (var kvp in perFileEntries)
-            {
-                string fileName = kvp.Key;
-                foreach (var fdrEntry in kvp.Value)
-                {
-                    // Prefer the 21-feature vector computed during coelution scoring.
-                    // Fall back to an all-zeros vector only for stub entries (e.g. loaded
-                    // from a Parquet cache without features) so the PercolatorEntry is
-                    // well-formed.
-                    double[] features;
-                    if (fdrEntry.Features != null &&
-                        fdrEntry.Features.Length == NUM_PIN_FEATURES)
-                    {
-                        features = fdrEntry.Features;
-                        nWithFeatures++;
-                    }
-                    else
-                    {
-                        features = BuildBasicFeatures(fdrEntry, libraryById);
-                        nWithoutFeatures++;
-                    }
-
-                    if (fdrEntry.IsDecoy)
-                        nInputDecoys++;
-                    else nInputTargets++;
-
-                    percEntries.Add(new PercolatorEntry
-                    {
-                        Id = string.Format("{0}_{1}", fileName, fdrEntry.EntryId),
-                        FileName = fileName,
-                        Peptide = fdrEntry.ModifiedSequence,
-                        Charge = fdrEntry.Charge,
-                        IsDecoy = fdrEntry.IsDecoy,
-                        EntryId = fdrEntry.EntryId,
-                        Features = features
-                    });
-                }
-            }
-
-            LogInfo(string.Format(
-                "[COUNT] Percolator input: {0} entries ({1} targets, {2} decoys, {3} features)",
-                percEntries.Count, nInputTargets, nInputDecoys, NUM_PIN_FEATURES));
-            LogInfo(string.Format(
-                "[COUNT] Percolator features computed: {0} entries with PIN features, {1} fallback",
-                nWithFeatures, nWithoutFeatures));
-
-            LogInfo(string.Format("Running Percolator on {0} entries...", percEntries.Count));
-
-            var percConfig = new PercolatorConfig
-            {
-                TrainFdr = config.RunFdr,
-                TestFdr = config.RunFdr,
-                MaxIterations = 10,
-                NFolds = 3,
-                FeatureNames = ParquetScoreCache.PIN_FEATURE_NAMES
-            };
-
-            // Streaming vs direct dispatch, matching Rust
-            // osprey/src/pipeline.rs::run_percolator_fdr. Above the
-            // MaxTrainSize * 2 threshold the training set is dominated by
-            // multi-observation-per-precursor redundancy; best-per-precursor
-            // dedup + peptide-grouped subsample give the SVM a diverse
-            // per-peptide training pool (same approach mokapot takes) and
-            // keep the Stage 5 standardizer fit on the subset -- essential
-            // for cross-impl byte parity with Rust once Astral-scale inputs
-            // push past the threshold.
-            PercolatorResults results;
-            if (percConfig.MaxTrainSize > 0 &&
-                percEntries.Count > percConfig.MaxTrainSize * 2)
-            {
-                results = RunPercolatorStreaming(percEntries, percConfig);
-            }
-            else
-            {
-                results = PercolatorFdr.RunPercolator(percEntries, percConfig);
-            }
-
-            // Map results back to FdrEntry stubs
-            var resultMap = new Dictionary<string, PercolatorResult>();
-            foreach (var result in results.Entries)
-                resultMap[result.Id] = result;
-
-            foreach (var kvp in perFileEntries)
-            {
-                string fileName = kvp.Key;
-                foreach (var fdrEntry in kvp.Value)
-                {
-                    string id = string.Format("{0}_{1}", fileName, fdrEntry.EntryId);
-                    PercolatorResult result;
-                    if (resultMap.TryGetValue(id, out result))
-                    {
-                        fdrEntry.Score = result.Score;
-                        fdrEntry.RunPrecursorQvalue = result.RunPrecursorQvalue;
-                        fdrEntry.RunPeptideQvalue = result.RunPeptideQvalue;
-                        fdrEntry.ExperimentPrecursorQvalue = result.ExperimentPrecursorQvalue;
-                        fdrEntry.ExperimentPeptideQvalue = result.ExperimentPeptideQvalue;
-                        fdrEntry.Pep = result.Pep;
-                    }
-                }
-            }
-
-            // Log FDR results
-            int nTargetPassing = 0;
-            int nDecoyPassing = 0;
-            foreach (var kvp in perFileEntries)
-            {
-                int fileTargets = 0;
-                int fileDecoys = 0;
-                foreach (var entry in kvp.Value)
-                {
-                    if (entry.EffectiveRunQvalue(config.FdrLevel) <= config.RunFdr)
-                    {
-                        if (entry.IsDecoy)
-                            fileDecoys++;
-                        else
-                            fileTargets++;
-                    }
-                }
-                LogInfo(string.Format(
-                    "[COUNT] Percolator pass [{0}]: {1} targets, {2} decoys at {3:P0} FDR",
-                    kvp.Key, fileTargets, fileDecoys, config.RunFdr));
-                nTargetPassing += fileTargets;
-                nDecoyPassing += fileDecoys;
-            }
-
-            LogInfo(string.Format(
-                "Percolator results: {0} targets, {1} decoys pass {2:P1} FDR",
-                nTargetPassing, nDecoyPassing, config.RunFdr));
-            LogInfo(string.Format(
-                "[COUNT] First-pass total across files: {0}",
-                nTargetPassing));
-
-            // Compute unique precursors across files (best q-value per modseq+charge)
-            var bestQByPrecursor = new Dictionary<string, double>(StringComparer.Ordinal);
-            foreach (var kvp in perFileEntries)
-            {
-                foreach (var entry in kvp.Value)
-                {
-                    if (entry.IsDecoy)
-                        continue;
-                    if (entry.EffectiveRunQvalue(config.FdrLevel) > config.RunFdr)
-                        continue;
-                    string pkey = entry.ModifiedSequence + "|" + entry.Charge;
-                    double q = entry.EffectiveRunQvalue(config.FdrLevel);
-                    double existing;
-                    if (!bestQByPrecursor.TryGetValue(pkey, out existing) || q < existing)
-                        bestQByPrecursor[pkey] = q;
-                }
-            }
-            LogInfo(string.Format(
-                "[COUNT] First-pass unique precursors (best q across files): {0}",
-                bestQByPrecursor.Count));
-        }
-
-        /// <summary>
-        /// Build a minimal PIN feature vector from an FdrEntry.
-        /// Used as a fallback ONLY when <see cref="FdrEntry.Features"/> has not been
-        /// populated (e.g. stubs loaded from a Parquet cache). In normal operation the
-        /// 21-feature vector is computed during coelution scoring in
-        /// <see cref="ScoreCandidate"/> and stored on the entry.
-        /// </summary>
-        private double[] BuildBasicFeatures(
-            FdrEntry entry, Dictionary<uint, LibraryEntry> libraryById)
-        {
-            double[] features = new double[NUM_PIN_FEATURES];
-
-            // 0: coelution_sum
-            features[0] = entry.CoelutionSum;
-            // 1: coelution_max (approximate as coelution_sum for basic version)
-            features[1] = entry.CoelutionSum * 0.5;
-            // 2: n_coeluting_fragments
-            features[2] = 3.0;
-            // 3: peak_apex
-            features[3] = 0.0;
-            // 4: peak_area
-            features[4] = 0.0;
-            // 5: peak_sharpness
-            features[5] = 0.0;
-            // 6: xcorr
-            features[6] = 0.0;
-            // 7: consecutive_ions
-            features[7] = 0.0;
-            // 8: explained_intensity
-            features[8] = 0.0;
-            // 9: mass_accuracy_mean
-            features[9] = 0.0;
-            // 10: abs_mass_accuracy_mean
-            features[10] = 0.0;
-            // 11: rt_deviation
-            features[11] = 0.0;
-            // 12: abs_rt_deviation
-            features[12] = 0.0;
-            // 13: ms1_precursor_coelution
-            features[13] = 0.0;
-            // 14: ms1_isotope_cosine
-            features[14] = 0.0;
-            // 15: median_polish_cosine
-            features[15] = 0.0;
-            // 16: median_polish_residual_ratio
-            features[16] = 0.0;
-            // 17: sg_weighted_xcorr
-            features[17] = 0.0;
-            // 18: sg_weighted_cosine
-            features[18] = 0.0;
-            // 19: median_polish_min_fragment_r2
-            features[19] = 0.0;
-            // 20: median_polish_residual_correlation
-            features[20] = 0.0;
-
-            return features;
-        }
-
-        /// <summary>
-        /// Streaming Percolator dispatch for multi-observation-per-precursor
-        /// inputs (total entries above <c>MaxTrainSize * 2</c>). Mirrors
-        /// Rust's <c>run_percolator_fdr</c> streaming branch
-        /// (osprey/src/pipeline.rs:4232-4580):
-        /// <list type="number">
-        /// <item>Best-per-precursor dedup across all per-file entries (one
-        /// target + one decoy per base_id, by Features[0] = coelution_sum).
-        /// </item>
-        /// <item>Peptide-grouped subsample to <c>MaxTrainSize</c> using the
-        /// same XOR-shift RNG and peptide-key sort order as Rust.</item>
-        /// <item>Train fold models + standardizer on that subset
-        /// (<c>TrainOnly = true</c>) -- the standardizer is fit on the
-        /// subset, matching Rust's run_percolator-on-subset behaviour
-        /// instead of fitting on the full 1M+ observation pool.</item>
-        /// <item>Apply the averaged model to ALL entries for scoring, then
-        /// compute PEP and per-run / experiment q-values on that flat
-        /// score array.</item>
-        /// </list>
-        /// The selection uses <see cref="PercolatorFdr.SelectBestPerPrecursor"/>
-        /// and <see cref="PercolatorFdr.SubsampleByPeptideGroup"/> -- the
-        /// same helpers the direct path calls internally, so both paths
-        /// select identical 300K subsets when given identical input.
-        /// </summary>
-        private PercolatorResults RunPercolatorStreaming(
-            List<PercolatorEntry> percEntries,
-            PercolatorConfig percConfig)
-        {
-            int n = percEntries.Count;
-            int maxTrain = percConfig.MaxTrainSize;
-
-            // Pull labels / entry IDs / peptides into flat arrays for the
-            // subset helpers.
-            var labels = new bool[n];
-            var entryIds = new uint[n];
-            var peptides = new string[n];
-            for (int i = 0; i < n; i++)
-            {
-                labels[i] = percEntries[i].IsDecoy;
-                entryIds[i] = percEntries[i].EntryId;
-                peptides[i] = percEntries[i].Peptide;
-            }
-
-            // 1. Best-per-precursor dedup.
-            int[] bestIdx = PercolatorFdr.SelectBestPerPrecursor(labels, entryIds, percEntries);
-            int dedupTargets = 0, dedupDecoys = 0;
-            for (int i = 0; i < bestIdx.Length; i++)
-            {
-                if (labels[bestIdx[i]]) dedupDecoys++;
-                else dedupTargets++;
-            }
-            LogInfo(string.Format(
-                "[COUNT] Percolator streaming best-per-precursor: {0} entries ({1} targets, {2} decoys) from {3} total",
-                bestIdx.Length, dedupTargets, dedupDecoys, n));
-
-            // 2. Peptide-grouped subsample if dedup count still exceeds MaxTrainSize.
-            int[] trainSubsetGlobalIdx;
-            if (maxTrain > 0 && bestIdx.Length > maxTrain)
-            {
-                var dedupLabels = new bool[bestIdx.Length];
-                var dedupEntryIds = new uint[bestIdx.Length];
-                var dedupPeptides = new string[bestIdx.Length];
-                for (int i = 0; i < bestIdx.Length; i++)
-                {
-                    int gi = bestIdx[i];
-                    dedupLabels[i] = labels[gi];
-                    dedupEntryIds[i] = entryIds[gi];
-                    dedupPeptides[i] = peptides[gi];
-                }
-                int[] localSelected = PercolatorFdr.SubsampleByPeptideGroup(
-                    dedupLabels, dedupEntryIds, dedupPeptides, maxTrain, percConfig.Seed);
-                trainSubsetGlobalIdx = new int[localSelected.Length];
-                for (int i = 0; i < localSelected.Length; i++)
-                    trainSubsetGlobalIdx[i] = bestIdx[localSelected[i]];
-            }
-            else
-            {
-                trainSubsetGlobalIdx = bestIdx;
-            }
-
-            int subTargets = 0, subDecoys = 0;
-            for (int i = 0; i < trainSubsetGlobalIdx.Length; i++)
-            {
-                if (labels[trainSubsetGlobalIdx[i]]) subDecoys++;
-                else subTargets++;
-            }
-            LogInfo(string.Format(
-                "[COUNT] Percolator streaming subsample: {0} entries ({1} targets, {2} decoys)",
-                trainSubsetGlobalIdx.Length, subTargets, subDecoys));
-
-            // 3. Build subset entry list + train.
-            var subsetEntries = new List<PercolatorEntry>(trainSubsetGlobalIdx.Length);
-            foreach (int i in trainSubsetGlobalIdx)
-                subsetEntries.Add(percEntries[i]);
-            var trainConfig = new PercolatorConfig
-            {
-                TrainFdr = percConfig.TrainFdr,
-                TestFdr = percConfig.TestFdr,
-                MaxIterations = percConfig.MaxIterations,
-                NFolds = percConfig.NFolds,
-                Seed = percConfig.Seed,
-                CValues = percConfig.CValues,
-                MaxTrainSize = percConfig.MaxTrainSize,
-                FeatureNames = percConfig.FeatureNames,
-                TrainOnly = true
-            };
-            PercolatorResults trainResults = PercolatorFdr.RunPercolator(subsetEntries, trainConfig);
-
-            // 4. Apply averaged model to ALL entries and compute q-values.
-            return PercolatorFdr.ScorePopulationAndComputeFdr(percEntries, trainResults, percConfig);
-        }
-
-        /// <summary>
-        /// Run simple target-decoy competition FDR (no machine learning).
-        /// Uses coelution_sum as the scoring function.
-        /// </summary>
-        private void RunSimpleFdr(
-            List<KeyValuePair<string, List<FdrEntry>>> perFileEntries,
-            OspreyConfig config)
-        {
-            var fdrController = new FdrController(config.RunFdr);
-
-            foreach (var kvp in perFileEntries)
-            {
-                var result = fdrController.CompeteAndFilter(
-                    kvp.Value,
-                    e => e.CoelutionSum,
-                    e => e.IsDecoy,
-                    e => e.EntryId);
-
-                LogInfo(string.Format(
-                    "  {0}: {1} targets pass (FDR={2:F4}, {3} target wins, {4} decoy wins)",
-                    kvp.Key, result.PassingTargets.Count, result.FdrAtThreshold,
-                    result.NTargetWins, result.NDecoyWins));
-
-                // Assign q-values based on simple competition
-                // Passing targets get fdr_at_threshold, non-passing get 1.0
-                var passingIds = new HashSet<uint>();
-                foreach (var target in result.PassingTargets)
-                    passingIds.Add(target.EntryId);
-
-                foreach (var entry in kvp.Value)
-                {
-                    if (!entry.IsDecoy && passingIds.Contains(entry.EntryId))
-                    {
-                        entry.RunPrecursorQvalue = result.FdrAtThreshold;
-                        entry.RunPeptideQvalue = result.FdrAtThreshold;
-                        entry.ExperimentPrecursorQvalue = result.FdrAtThreshold;
-                        entry.ExperimentPeptideQvalue = result.FdrAtThreshold;
-                    }
-                }
-            }
-        }
-
         #endregion
 
         #region Stage 8: Protein FDR
-
-        /// <summary>
-        /// First-pass protein FDR run BEFORE Stage 6 reconciliation, on the
-        /// full pre-compaction peptide pool. Sets only RunProteinQvalue
-        /// (leaves ExperimentProteinQvalue at its 1.0 default for the
-        /// second-pass to overwrite). Detected-peptide filter uses
-        /// run_peptide_qvalue, the strict peptide-level gate, matching Rust
-        /// pipeline.rs:3045-3049 exactly. Protein-FDR gate is config.RunFdr
-        /// (1x), the Savitski-2015 convention applied at first pass, NOT the
-        /// 2x relaxed gate the post-output Stage 8 protein FDR uses.
-        /// </summary>
-        internal void RunFirstPassProteinFdr(
-            List<KeyValuePair<string, List<FdrEntry>>> perFileEntries,
-            List<LibraryEntry> fullLibrary,
-            OspreyConfig config)
-        {
-            // Detected-peptide gate: targets passing peptide-level run FDR.
-            // Matches Rust pipeline.rs:3048 (e.run_peptide_qvalue <= run_fdr).
-            var detectedPeptides = new HashSet<string>(StringComparer.Ordinal);
-            foreach (var kvp in perFileEntries)
-            {
-                foreach (var entry in kvp.Value)
-                {
-                    if (!entry.IsDecoy && entry.RunPeptideQvalue <= config.RunFdr)
-                        detectedPeptides.Add(entry.ModifiedSequence);
-                }
-            }
-            LogInfo(string.Format(
-                "[COUNT] First-pass detected peptides for protein FDR: {0} unique",
-                detectedPeptides.Count));
-
-            var parsimony = ProteinFdr.BuildProteinParsimony(
-                fullLibrary, config.SharedPeptides, detectedPeptides);
-
-            // Best peptide score across all files for picked-protein TDC.
-            var bestScores = ProteinFdr.CollectBestPeptideScores(perFileEntries);
-
-            // First-pass gate is config.RunFdr exactly — matches Rust
-            // pipeline.rs:3062 (compute_protein_fdr at config.run_fdr).
-            var proteinFdr = ProteinFdr.ComputeProteinFdr(parsimony, bestScores, config.RunFdr);
-
-            int nAtRunFdr = 0;
-            foreach (var qv in proteinFdr.GroupQvalues.Values)
-            {
-                if (qv <= config.RunFdr)
-                    nAtRunFdr++;
-            }
-            LogInfo(string.Format(
-                "First-pass protein FDR: {0} target groups at {1:P1} FDR",
-                nAtRunFdr, config.RunFdr));
-
-            if (OspreyDiagnostics.DumpProteinFdr)
-            {
-                OspreyDiagnostics.WriteStage6ProteinFdrDump(
-                    bestScores, proteinFdr.PeptideQvalues);
-                if (OspreyDiagnostics.ProteinFdrOnly)
-                    OspreyDiagnostics.ExitAfterDump(@"OSPREY_PROTEIN_FDR_ONLY");
-            }
-
-            // Set RunProteinQvalue ONLY. Experiment-protein-q is set by the
-            // post-output Stage 8 protein FDR pass (Rust calls it second-pass).
-            ProteinFdr.PropagateProteinQvalues(perFileEntries, proteinFdr,
-                setRun: true, setExperiment: false);
-        }
 
         #endregion
